fix(api): move TypedDicts to dev_types for type checking only
